backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.23.2
pyyaml>=6.0
